+++ NEWS.md
@@ -1,6 +1,6 @@
 # inlmisc 0.4.0.9000
 
-- In LaTeX "misc/preamble.tex" file, fix spacing after number in List of Figures and List of Tables.
+- Tidy LaTeX "misc/preamble.tex" file and include fix for spacing after number in List of Figures and List of Tables.
 
 - In `AddInsetMap`, add "center" option for `loc` argument.